refactor: extract utils from OrderRow
