+++ apps/cowswap-frontend/src/modules/ordersTable/pure/OrdersTableContainer/OrderRow/index.tsx
@@ -3,11 +3,11 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import orderPresignaturePending from '@cowprotocol/assets/cow-swap/order-presignature-pending.svg'
 import { ZERO_FRACTION } from '@cowprotocol/common-const'
 import { useTimeAgo } from '@cowprotocol/common-hooks'
-import { formatDateWithTimezone, getAddress, getEtherscanLink } from '@cowprotocol/common-utils'
+import { formatDateWithTimezone, getAddress } from '@cowprotocol/common-utils'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { TokenLogo } from '@cowprotocol/tokens'
 import { Command, UiOrderType } from '@cowprotocol/types'
-import { HoverTooltip, Loader, PercentDisplay, percentIsAlmostHundred, TokenAmount, UI } from '@cowprotocol/ui'
+import { HoverTooltip, Loader, PercentDisplay, percentIsAlmostHundred, TokenAmount } from '@cowprotocol/ui'
 import { useIsSafeWallet } from '@cowprotocol/wallet'
 import { Currency, CurrencyAmount, Percent, Price } from '@uniswap/sdk-core'
 
@@ -22,18 +22,13 @@ import { PendingOrderPrices } from 'modules/orders/state/pendingOrdersPricesAtom
 import { getIsEthFlowOrder } from 'modules/swap/containers/EthFlowStepper'
 import { BalancesAndAllowances } from 'modules/tokens'
 
-import {
-  FAIR_PRICE_THRESHOLD_PERCENTAGE,
-  GOOD_PRICE_THRESHOLD_PERCENTAGE,
-  PENDING_EXECUTION_THRESHOLD_PERCENTAGE,
-} from 'common/constants/common'
+import { PENDING_EXECUTION_THRESHOLD_PERCENTAGE } from 'common/constants/common'
 import { useSafeMemo } from 'common/hooks/useSafeMemo'
 import { RateInfo } from 'common/pure/RateInfo'
 import { getQuoteCurrency } from 'common/services/getQuoteCurrency'
 import { isOrderCancellable } from 'common/utils/isOrderCancellable'
 import { calculatePercentageInRelationToReference } from 'utils/orderUtils/calculatePercentageInRelationToReference'
 import { calculatePriceDifference, PriceDifference } from 'utils/orderUtils/calculatePriceDifference'
-import { getIsComposableCowParentOrder } from 'utils/orderUtils/getIsComposableCowParentOrder'
 import { getIsFinalizedOrder } from 'utils/orderUtils/getIsFinalizedOrder'
 import { getSellAmountWithFee } from 'utils/orderUtils/getSellAmountWithFee'
 import { getUiOrderType } from 'utils/orderUtils/getUiOrderType'
@@ -43,28 +38,15 @@ import { EstimatedExecutionPrice } from './EstimatedExecutionPrice'
 import { OrderContextMenu } from './OrderContextMenu'
 import { WarningTooltip } from './OrderWarning'
 import * as styledEl from './styled'
+import { getActivityUrl, getDistanceColor, shouldShowDashForExpiration } from './utils'
 
-import { OrderParams } from '../../../utils/getOrderParams'
-import { getOrderParams } from '../../../utils/getOrderParams'
+import { getOrderParams, OrderParams } from '../../../utils/getOrderParams'
 import { OrderStatusBox } from '../../OrderStatusBox'
 import { CheckboxCheckmark, TableRow, TableRowCheckbox, TableRowCheckboxWrapper } from '../styled'
 import { OrderActions } from '../types'
 
 // Constants
 const TIME_AGO_UPDATE_INTERVAL = 3000
-
-// Helper to determine the color based on percentage
-function getDistanceColor(percentage: number): string {
-  const absPercentage = Math.abs(percentage)
-
-  if (absPercentage <= GOOD_PRICE_THRESHOLD_PERCENTAGE) {
-    return `var(${UI.COLOR_SUCCESS})` // Green - good price
-  } else if (absPercentage <= FAIR_PRICE_THRESHOLD_PERCENTAGE) {
-    return `var(${UI.COLOR_PRIMARY})` // Blue - fair price
-  }
-
-  return 'inherit' // Default text color for larger differences
-}
 
 function CurrencyAmountItem({ amount }: { amount: CurrencyAmount<Currency> }) {
   return (
@@ -902,47 +884,4 @@ function useFeeAmountDifference(
     () => calculatePercentageInRelationToReference({ value: feeAmount, reference: inputCurrencyAmount }),
     [feeAmount, inputCurrencyAmount],
   )
-}
-
-function getActivityUrl(chainId: SupportedChainId, order: ParsedOrder): string | undefined {
-  const { activityId } = order.executionData
-
-  if (getIsComposableCowParentOrder(order)) {
-    return undefined
-  }
-
-  if (order.composableCowInfo?.isVirtualPart) {
-    return undefined
-  }
-
-  if (order.status === OrderStatus.SCHEDULED) {
-    return undefined
-  }
-
-  return chainId && activityId ? getEtherscanLink(chainId, 'transaction', activityId) : undefined
-}
-
-function shouldShowDashForExpiration(order: ParsedOrder): boolean {
-  // Show dash for finalized orders that are not expired
-  if (getIsFinalizedOrder(order) && order.status !== OrderStatus.EXPIRED) {
-    return true
-  }
-
-  // For TWAP parent orders, show dash when all child orders are in a final state
-  if (getIsComposableCowParentOrder(order)) {
-    // If the parent order is fulfilled or cancelled, all child orders are finalized
-    if (order.status === OrderStatus.FULFILLED || order.status === OrderStatus.CANCELLED) {
-      return true
-    }
-
-    // For mixed states (some filled, some expired), check either condition:
-    // 1. fullyFilled: true when all non-expired parts are filled
-    // 2. status === EXPIRED: true when all remaining parts are expired
-    // Either condition indicates all child orders are in a final state
-    if (order.executionData.fullyFilled || order.status === OrderStatus.EXPIRED) {
-      return true
-    }
-  }
-
-  return false
 }
